Implement loading popup for download operations
Added a loading popup component that shows during file download processes in FileSharing.tsx and Columns.tsx. Also included a new CircleLoader component for the load animation. This popup set to display when a download operation begins and to desist after the operation is completed. This helps in better user experience by indicating that a process is ongoing.