Improve file upload UI in UploadItemDialog
Reworked the file uploader in UploadItemDialog to be a more interactive drop zone for files. Updated DropZone utility component to display chosen files in a list, added a limit to restrict upload upto only 5 items at a time and improved file type handling. Also, the color of the upload button was changed for better visibility.