Refactor WebDav FileManager to functional structure
In this commit the previously class-based WebDav FileManager is converted to functional structure. Extraction of some functionalities into reusable utility functions improves reusability. Updates are also made in several components to correctly import and use these functions. This refactor enhances code readability, maintainability and aligns with the rest of the functional structure throughout the project.