Integrate OnlyOffice document editor and refactor FileSharing component
This commit incorporates the OnlyOffice document editor into our application and refactors the FileSharing component's menu for a more dynamic user experience. It also includes the addition of the 'useMenuItems' hook, which fetches and prepares menu items. The useMenuBarConfig hook was also updated to use the newly added 'useMenuItems' and properly configure menu bar according to current location. Finally, the commit cleans up commented lines and simplifies the HexagonButton import.