Improve collapsibility behavior of the menu bar
Added a functionality to make the vertical menu bar collapsible and improved its handling on mobile screens. Underlying changes include the addition of an icon for toggle, adjustments in element visibility and size when the bar is collapsed. Styling transitions have also been added to smooth out this collapsing behavior, providing a cleaner user interface experience.