Refactor translation implementation
Transitioned translation function across multiple files from using 'useTranslation' hook to using 'translateKey' utility function. Also, corrected small typos in the locale files. This refactor abstracts the translation logic away from the component, achieving a cleaner code and making it easier to apply changes to the translation system in the future.

All in all it should resolve the recommended changes in the MR