The provided commit message does not correspond to the diff. Here's a relevant message:
Implement filename truncation in file sharing table

In the file sharing table, a truncate function has been added to limit the length of a filename to a specified number of characters (default of 10). If the filename exceeds this limit, it will be cut off and followed by an ellipsis. This prevents overly long filenames from disrupting the layout or readability of the table.