Refactor dialog open prop from isDialogOpen/isSheetOpen to isOpen
The naming of the prop that controlled whether the SettingsDialog or Sheet was open has been simplified. Previously, we had two props: isDialogOpen and isSheetOpen. This was unnecessarily complicated and has been merged into a single prop named isOpen. This change was made in all files where these props were used. The name 'isOpen' more clearly signifies its purpose and reduces confusion. Additionally, a localization fix was made for the 'Close' string in Sheet, and a new ENV variable 'VITE_ROLE' was added.