Modify translation keys in WebDavFileManager and i18n files
This commit updates the translation keys in the WebDavFileManager and i18n files. It includes changes to properly outline the syntax for the translation parameters and fixes how the destination path is passed to the move_successful key. Also, debug mode has been activated in i18n.ts to facilitate debugging.