Add support for previewing and editing docx and PDF files
This refactoring prepares the app for editing docx and PDF files by opening them in a new '/preview' route. It introduces a new state in the FileEditor store for managing files that are to be edited. Updates were also made in the backend to provide a route for file download linked to the onlyoffice DocumentServer via a new WebDavController.