Add folder creation functionality in file manager service
This commit introduces a new method for creating folders in the file manager service. The 'createFolder' function has been added to both the service and controller files of the filemanager, using the 'MKCOL' method in the webdav service for implementation. This allows users to create new folders at a specified path of choice.