Refactor file sharing module and improve UI
Significantly refactored the structure of the file sharing module by removing unnecessary files, moving some files to shared libraries and renaming others for better clarity. An improvement was made on the UI by applying a new configuration for displayable file types. Additionally, a new service was introduced for handling file sharing configuration on the server-side.