Add editing mode and routing for file viewer
This commit introduces an editing mode for the file viewer component and adds related routes accordingly. Moreover, changes include the implementation of more specific error handling for the OnlyOfficeEditor component, and forcesave set to true for OnlyOffice configurations. Also, the addition of a new 'EmptyLayout' component and changes to the FilePreviewOptionsBar button functionality are included.